Clarify Giffer modes (yes, "mode", not "status")
I found it really difficult to reason about what was showing when.

For the controls, we can render only the ones that we're showing at a
given time. For the preview and the video stream, we cannot. For various
functionality here to work (all the `this.video` and `this.preview`
stuff), we need these to always be on the page.

Why not use routes?
-------------------

At first, I attempted to use routes instead of using component state,
for displaying the correct navigational buttons. It almost looks like a
routing structure already, after all.

However, Giffer is a poor candidate for routes. We are doing some very
procedural work, based on user input. Using component state works better
for this sor of procedural logic. Also, we don't really want a route
that means "start shooting, immediately."