Make App in charge of fetching & passing data
This is a big change.

Rather than having data fetched from each individual component, this
fetches it at the App level. Then App's routes are rendered differently,
and data is passed into each one.

This in _theory_ allows us to add nice transitions between the list of
all GIFs and an individual GIF's page. In practice, there's still a bit
of lag... presumably because the GIF data is so huge. This seems like
another reason to use videos instead of GIFs (see [comment here][1]).

  [1]: https://github.com/citrusbyte/thicket/pull/11#discussion_r141730036