Extract NavLinks/BottomNav to thicket-elements
Things to note:

1. The inclusion of `Link` is a little bit goofy - in fact, if we try to
use that version in the `thicket-elements` index page (the one loaded by
`yarn start` within the `thicket-elements` directory), we get an error.
However, this seemed like the fastest way to extract this library and
make it interop with the current webapp.

2. I have renamed this file to `BottomNav`, as that is what it mostly
encapsulates. Whether they're buttons or links, most of the components
in this file (and all the ones that will be shared between the camera
component and the main component) are _bottom_ navigational items.

3.  The `Back` link has been removed from this new `BottomNav` file.
Though this back link is a _nav_ link, it is not _bottom_ nav like the
rest of the components.