Dragon post: Chad's edit
This is a big edit, but in the end I decided to commit it right to the
same branch, rather than creating a PR back to this one.

@AquiGorka, you can of course feel free to change anything as you see
fit. Put it back how it was, if you want.

I attempted to keep the flow mostly the same, but to clarify various
topics and correct some factual errors. I hope it doesn't sound too much
like my voice now, but instead sounds how you would want it to sound,
Gorka.