[FAILED] Intro: try to move loader to top level
This attempts to make the `sticky` styling work better by moving the
`GifToEarthProgress` component to the top level.

Unfortunately, this breaks the styling of the component when it should
be part of the flow of the page--the part before its sticky. I don't see
a good way around this problem, if I'm going to stick with `position:
sticky`.