Intro: affix header when needed
This uses `requestAnimationFrame` as encuraged by MDN [1] to avoid
scroll jank [2].

It required quite a few changes to the styling of the box around the
loader. I had to make it look like it was part of the page flow before
making it `position: fixed`, but it actually works best to completely
remove it from the page flow all along, so the content below it doesn't
jump around when its positioning changes. That's why it is `position:
absolute` by default, and why the element after it is given a margin-top
equivalent to its height. It's also why various measurements were
changed to pixels instead of em, so that its height can be calculated
more precisely.

This is frustratingly brittle--if we change styles elsewhere on the page
that seem to be unrelated, like the amount of padding, then this will
require parallel changes. We might not notice that it requires such
changes right away, which could result in styling bugs that only affect
certain screen resolutions.

  [1]: https://developer.mozilla.org/en-US/docs/Web/Events/scroll#Example
  [2]: https://developers.google.com/web/updates/2016/06/passive-event-listeners?hl=en