Webapp: Set up s3-deploy
In order for the deploy script to work, you need to have a
`~/.aws/credentials` file set up as explained [in the `s3-deploy`
readme][1].

To do that, you will need to have a DevOps person at CB set up your
account with proper permissions, and you will need to get your access
key and access key secret.

This currently uploads files to a Citrusbyte S3 bucket, which is currently
served only at the insecure URL http://thicket.citrusbyte.com.s3-website-us-west-2.amazonaws.com/welcome
Since it's an insecure URL, basically nothing works. We can't access the
camera, we can't use websockets. We need a CB DevOps person to help us
set up a Route53 endpoint with TLS in order to really start using this
approach.

Also, we don't want this deploy script to be part of the open source
repo. I need to move this to `thicket-tm`.

  [1]: https://www.npmjs.com/package/s3-deploy#aws-credentials