Store: add initial subscribe to communities & publications
This change makes the store subscribe to communities and publications
even if the user has not visited the sections.
Pros: updates will be received and cached "in the background", when the
user visits the related sections that data will already be in cache,
thus the app will work "faster".
Cons: an initial sync for a lot of data can make the app unusable. As it
is, this change is asynchronous but the side effects on a large data
synch should be optimized.