Intro: Hide banner after clicking to it
As Sam mocked up in #80, it makes sense to hide the "interplanetary GIFs
exchanged" message after it's no longer relevant. While not a perfect
solution, a quick way to do this is to add an `onClick` handler to the
"Check it out!" link.

This will cause the banner to stick around if the user scrolls to the
section, rather than clicking the link. This doesn't seem like a big
enough problem for me to exert the effort to fix it.