Intro: try to customize button hover text color
This approach doesn't currently work in Chrome, but I think it might at
some point?

Not a huge deal that it doesn't work. It would have been nice to have
that text be red, but it didn't seem worth more than a half hour of my
time to try to make it work.