App: setup app layout
This app will have a fixed header, its container adds the header's
height as top padding. This way the content inside will not be hidden
below the header until it is scrolled.

The app's main container is #app even though this element is child to
two other elements: body and #root.
React discourages using body as the main container.
React router dom needs its children to come from one element so #root
could not be used as the main container either.

For the above reasons html, body and #root have an overflow property of
hidden and #app has it set to auto.