Communitites: add responsive grid
This changes make the grid act as one column for mobile and as a
"wrap-able" grid for desktop. Meaning it will allow as many elements as
possible (that fit into the screen width) and then wrap.
The last row will align to the left.

The header uses the same class as the list grid so that it aligns
according to the grids width, thus the title will be aligned to the
grid's first element.