Fix elements images locations
While storing the images within the directory that uses them works
correctly from a `yarn start` inside the `thicket-elements` directory,
it does not work correctly once the package is built.

I would love to find a better way to do this, but the only way that I've
found so far is to put the images in the `images` directory.