Webapp/App: Make tags semantic; rm profile link
In the new designs, the link to the profile is no longer present on
every single page. We need to think through how to get it only on the
pages we want. But for now, to make the new Welcome page look good, I'm
removing it altogether.

Also, while I was snooping around in this file, I found a few tags that
seem inappropriate.

* Rather than `<main>` wrapping all content, including the header, I
  think it makes more sense to have it wrap only the main content.
* Rather than using a `<div>` for the header, it's straightforward to
  use a `<header>`