Syncing: add listen to syncing & synced events
When syncing show labels on empty Community title and Community size.
No need to listen to peeer left & joined events when yjs has on userEvent.
Syncing can come in any time, with this changes we assume the user is
alone in the room, and when any other user joins then the syncing
begins.
With this changes the GIFs will show a placeholder and then render the
GIFs.